+++ plugin-dev/Source/Sentry/Private/Android/Java/SentryBridgeJava.java
@@ -3,8 +3,6 @@
 package io.sentry.unreal;
 
 import android.app.Activity;
-import android.content.pm.PackageInfo;
-import android.content.pm.PackageManager;
 
 import androidx.annotation.NonNull;
 
@@ -13,6 +11,7 @@ import java.util.Map;
 
 import io.sentry.Attachment;
 import io.sentry.Breadcrumb;
+import io.sentry.Hint;
 import io.sentry.IHub;
 import io.sentry.Scope;
 import io.sentry.ScopeCallback;
@@ -33,6 +32,7 @@ public class SentryBridgeJava {
 			final String releaseName,
 			final String environment,
 			final String gameLogPath,
+			final String gameBackupLogPath,
 			final boolean enableAutoSessionTracking,
 			final long sessionTimeout,
 			final boolean enableStackTrace) {
@@ -45,15 +45,24 @@ public class SentryBridgeJava {
 				options.setEnableAutoSessionTracking(enableAutoSessionTracking);
 				options.setSessionTrackingIntervalMillis(sessionTimeout);
 				options.setAttachStacktrace(enableStackTrace);
-			}
-		});
-
-		Sentry.configureScope(new ScopeCallback() {
-			@Override
-			public void run(@NonNull Scope scope) {
-				if(!gameLogPath.isEmpty()) {
-					scope.addAttachment(new Attachment(gameLogPath));
-				}
+				options.setBeforeSend(new SentryOptions.BeforeSendCallback() {
+					@Override
+					public SentryEvent execute(SentryEvent event, Hint hint) {
+						if(event.isCrashed() && event.isErrored())
+						{
+							if(!gameBackupLogPath.isEmpty()) {
+								hint.addAttachment(new Attachment(gameBackupLogPath));
+							}
+						}
+						else
+						{
+							if(!gameLogPath.isEmpty()) {
+								hint.addAttachment(new Attachment(gameLogPath));
+							}
+						}
+						return event;
+					}
+				});
 			}
 		});
 	}