+++ plugin-dev/Source/Sentry/Private/Utils/SentryFileUtils.h
@@ -1,0 +1,12 @@
+// Copyright (c) 2023 Sentry. All Rights Reserved.
+
+#pragma once
+
+#include "CoreMinimal.h"
+
+class SentryFileUtils
+{
+public:
+	static FString GetGameLogPath();
+	static FString GetGameLogBackupPath();
+};